Rename MAVSDK-CSharp to MAVSDK
