+++ MAVSDK-CSharp/MAVSDK/MAVSDK.csproj
