+++ MAVSDK-CSharp/MAVSDK/Plugins/Calibration.cs
