+++ MAVSDK-CSharp/MAVSDK/Plugins/Camera.cs
