+++ MAVSDK-CSharp/MAVSDK/Plugins/Core.cs
