+++ MAVSDK-CSharp/MAVSDK/Plugins/Gimbal.cs
