+++ MAVSDK-CSharp/MAVSDK/templates/call.j2
