+++ MAVSDK-CSharp/MAVSDK/templates/enum.j2
