+++ MAVSDK-CSharp/MAVSDK/templates/file.j2
@@ -10,7 +10,7 @@ using Mavsdk.Rpc.{{ plugin_name.upper_camel_case }};
 
 using Version = Mavsdk.Rpc.Info.Version;
 
-namespace MAVSDK_CSharp.Plugins
+namespace MAVSDK.Plugins
 {
     public class {{ plugin_name.upper_camel_case }}
     {
