+++ MAVSDK-CSharp/MAVSDK/templates/request.j2
