+++ MAVSDK-CSharp/MAVSDK/templates/stream.j2
