+++ MAVSDK-CSharp/MAVSDK/templates/struct.j2
