+++ MAVSDK-CSharp/MAVSDK/templates/type_conversions
